requirements.txt
django
django-crispy-forms
django-bootstrap3
django-datetimepicker
